"codes/calculadora Python/funciones/calcular.py":

+++ codes/calculadora Python/funciones/calcular.py
@@ -1,22 +1,22 @@
 
 def sumar():
-    a = int(input("ingrese el valor para el numero 1:"))
-    b = int(input("ingrese el valor para el numero 2:"))
+    a = float(input("ingrese el valor para el numero 1:"))
+    b = float(input("ingrese el valor para el numero 2:"))
     return(a+b)
 
 def restar():
-    a = int(input("ingrese el valor para el numero 1:"))
-    b = int(input("ingrese el valor para el numero 2:"))
+    a = float(input("ingrese el valor para el numero 1:"))
+    b = float(input("ingrese el valor para el numero 2:"))
     return(a-b)
 
 def multiplicar():
-    a = int(input("ingrese el valor para el numero 1:"))
-    b = int(input("ingrese el valor para el numero 2:"))
+    a = float(input("ingrese el valor para el numero 1:"))
+    b = float(input("ingrese el valor para el numero 2:"))
     return(a*b)
 
 def dividir():
-    a = int(input("ingrese el valor para el numero 1:"))
-    b = int(input("ingrese el valor para el numero 2:"))
+    a = float(input("ingrese el valor para el numero 1:"))
+    b = float(input("ingrese el valor para el numero 2:"))
     return(a/b)
 
 def limpiar():
